:sparkles:  All requested fixes done
Removed from index.js and from types.d.ts the function isAdFinished since there is an event to manage if the Ad has finished.
In General use.md: Replaced if (ct.gamedistribution.adPlaying) { to if (!ct.gamedistribution.adPlaying) {
In module.js Replaced :
"name": "Ad is finished",
"parameterizedName": "Check if an Ad is finished.",
to
"name": "Ad has finished",
"parameterizedName": "Check if an Ad has finished.",

Removed the parametrizedName sin I didn't know there is no need to use it if there is no parameters to display.

Added .vscode to .gitignore so is not included the pull.